# Don't parse string imports if not enabled