--- a/src/python/pants/backend/python/dependency_inference/import_parser.py
+++ b/src/python/pants/backend/python/dependency_inference/import_parser.py
@@ -7,13 +7,12 @@ from pants.backend.python.util_rules.pex import PexInterpreterConstraints
 from pants.backend.python.util_rules.pex_environment import PythonExecutable
 from pants.core.util_rules.source_files import SourceFilesRequest
 from pants.core.util_rules.stripped_source_files import StrippedSourceFiles
+from pants.engine.collection import DeduplicatedCollection
 from pants.engine.fs import CreateDigest, Digest, FileContent, MergeDigests
 from pants.engine.process import Process, ProcessResult
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
 from pants.engine.target import Sources
 from pants.util.logging import LogLevel
-from pants.util.memo import memoized_property
-from pants.util.ordered_set import FrozenOrderedSet
 
 _SCRIPT = """\
 # -*- coding: utf-8 -*-
@@ -24,7 +23,7 @@ from __future__ import print_function, unicode_literals
 
 from io import open
 import ast
-import os.path
+import os
 import re
 import sys
 
@@ -35,16 +34,15 @@ STRING_IMPORT_REGEX = re.compile(r"^([a-z_][a-z_\\d]*\\.){2,}[a-zA-Z_]\\w*$", re
 class AstVisitor(ast.NodeVisitor):
     def __init__(self, package_parts):
         self._package_parts = package_parts
-        self.explicit_imports = set()
-        self.string_imports = set()
+        self.imports = set()
 
     def maybe_add_string_import(self, s):
         if STRING_IMPORT_REGEX.match(s):
-            self.string_imports.add(s)
+            self.imports.add(s)
 
     def visit_Import(self, node):
         for alias in node.names:
-            self.explicit_imports.add(alias.name)
+            self.imports.add(alias.name)
 
     def visit_ImportFrom(self, node):
         if node.level:
@@ -57,7 +55,7 @@ class AstVisitor(ast.NodeVisitor):
         else:
             abs_module = node.module
         for alias in node.names:
-            self.explicit_imports.add("{}.{}".format(abs_module, alias.name))
+            self.imports.add("{}.{}".format(abs_module, alias.name))
 
     def visit_Call(self, node):
       # Handle __import__("string_literal").  This is commonly used in __init__.py files,
@@ -71,37 +69,38 @@ class AstVisitor(ast.NodeVisitor):
           if sys.version_info[0:2] < (3, 8) and isinstance(node.args[0], ast.Str):
               arg_s = node.args[0].s
               val = arg_s.decode("utf-8") if isinstance(arg_s, bytes) else arg_s
-              self.explicit_imports.add(arg_s)
+              self.imports.add(arg_s)
               return
           elif isinstance(node.args[0], ast.Constant):
-              self.explicit_imports.add(str(node.args[0].value))
+              self.imports.add(str(node.args[0].value))
               return
       self.generic_visit(node)
 
-# String handling changes a bit depending on Python version. We dynamically add the appropriate
-# logic.
-if sys.version_info[0:2] == (2,7):
-    def visit_Str(self, node):
-        try:
-            val = node.s.decode("utf8") if isinstance(node.s, bytes) else node.s
-            self.maybe_add_string_import(val)
-        except UnicodeError:
-            pass
+if os.environ["STRING_IMPORTS"] == "y":
+    # String handling changes a bit depending on Python version. We dynamically add the appropriate
+    # logic.
+    if sys.version_info[0:2] == (2,7):
+        def visit_Str(self, node):
+            try:
+                val = node.s.decode("utf8") if isinstance(node.s, bytes) else node.s
+                self.maybe_add_string_import(val)
+            except UnicodeError:
+                pass
 
-    setattr(AstVisitor, 'visit_Str', visit_Str)
+        setattr(AstVisitor, 'visit_Str', visit_Str)
 
-elif sys.version_info[0:2] < (3, 8):
-    def visit_Str(self, node):
-        self.maybe_add_string_import(node.s)
+    elif sys.version_info[0:2] < (3, 8):
+        def visit_Str(self, node):
+            self.maybe_add_string_import(node.s)
 
-    setattr(AstVisitor, 'visit_Str', visit_Str)
+        setattr(AstVisitor, 'visit_Str', visit_Str)
 
-else:
-    def visit_Constant(self, node):
-        if isinstance(node.value, str):
-            self.maybe_add_string_import(node.value)
+    else:
+        def visit_Constant(self, node):
+            if isinstance(node.value, str):
+                self.maybe_add_string_import(node.value)
 
-    setattr(AstVisitor, 'visit_Constant', visit_Constant)
+        setattr(AstVisitor, 'visit_Constant', visit_Constant)
 
 
 def parse_file(filename):
@@ -114,8 +113,7 @@ def parse_file(filename):
 
 
 if __name__ == "__main__":
-    explicit_imports = set()
-    string_imports = set()
+    imports = set()
 
     for filename in sys.argv[1:]:
         tree = parse_file(filename)
@@ -124,41 +122,29 @@ if __name__ == "__main__":
 
         package_parts = os.path.dirname(filename).split(os.path.sep)
         visitor = AstVisitor(package_parts)
-        visitor.visit(tree)
 
-        explicit_imports.update(visitor.explicit_imports)
-        string_imports.update(visitor.string_imports)
+        visitor.visit(tree)
+        imports.update(visitor.imports)
 
     # We have to be careful to set the encoding explicitly and write raw bytes ourselves.
     # See below for where we explicitly decode.
-    buffer = sys.stdout if sys.version_info[0:2] == (2,7) else sys.stdout.buffer
-    buffer.write("\\n".join(sorted(explicit_imports)).encode("utf8"))
-    buffer.write(b"\\n--\\n")
-    buffer.write("\\n".join(sorted(string_imports)).encode("utf8"))
+    buffer = sys.stdout if sys.version_info[0:2] == (2, 7) else sys.stdout.buffer
+    buffer.write("\\n".join(sorted(imports)).encode("utf8"))
 """
 
 
-@dataclass(frozen=True)
-class ParsedPythonImports:
+class ParsedPythonImports(DeduplicatedCollection[str]):
     """All the discovered imports from a Python source file.
 
-    Explicit imports are imports from `import x` and `from module import x` statements. String
-    imports come from strings that look like module names, such as
-    `importlib.import_module("example.subdir.Foo")`.
+    May include string imports if the request specified to include them.
     """
-
-    explicit_imports: FrozenOrderedSet[str]
-    string_imports: FrozenOrderedSet[str]
-
-    @memoized_property
-    def all_imports(self) -> FrozenOrderedSet[str]:
-        return FrozenOrderedSet(sorted([*self.explicit_imports, *self.string_imports]))
 
 
 @dataclass(frozen=True)
 class ParsePythonImportsRequest:
     sources: Sources
     interpreter_constraints: PexInterpreterConstraints
+    string_imports: bool
 
 
 @rule
@@ -181,16 +167,13 @@ async def parse_python_imports(request: ParsePythonImportsRequest) -> ParsedPyth
             ],
             input_digest=input_digest,
             description=f"Determine Python imports for {request.sources.address}",
+            env={"STRING_IMPORTS": "y" if request.string_imports else "n"},
             level=LogLevel.DEBUG,
         ),
     )
     # See above for where we explicitly encoded as utf8. Even though utf8 is the
     # default for decode(), we make that explicit here for emphasis.
-    explicit_imports, _, string_imports = process_result.stdout.decode("utf8").partition("--")
-    return ParsedPythonImports(
-        explicit_imports=FrozenOrderedSet(explicit_imports.strip().splitlines()),
-        string_imports=FrozenOrderedSet(string_imports.strip().splitlines()),
-    )
+    return ParsedPythonImports(process_result.stdout.decode("utf8").strip().splitlines())
 
 
 def rules():