We should not be doing wasted work. If string imports won't be consumed, no need to waste time traversing the AST for these imports + decoding/encoding the detected imports.

Beyond performance, this simplifies some of the code.

### Benchmarks
Runs `multitime -n 20 ./pants --no-pantsd --no-process-execution-local-cache dependencies src/python::` 20 times.

Before:

```            
            Mean        Std.Dev.    Min         Median      Max
real        4.220       0.238       4.047       4.143       5.119
user        6.446       0.212       6.221       6.381       7.138
sys         2.916       0.081       2.783       2.903       3.155
```


After:

```
            Mean        Std.Dev.    Min         Median      Max
real        4.180       0.163       3.899       4.158       4.548
user        6.407       0.136       6.264       6.371       6.757
sys         2.856       0.061       2.710       2.855       2.991
```

Even though this benchmark does not show a huge gain, this PR does result in doing less work so it makes sense that there is a minor improvement. 

[ci skip-rust]
[ci skip-build-wheels]